% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ucsc_course_skeleton.R
\name{create_book}
\alias{create_book}
\alias{create_bs4_ucsc_course}
\title{Create a bs4 bookdown project for UCSC course}
\usage{
create_bs4_ucsc_course(path)
}
\arguments{
\item{path}{Absolute path to an empty directory in which to create the bookdown project.
In the RStudio IDE, if \pkg{rstudioapi} package available,
an RStudio project will be created.}
}
\description{
Create a bookdown project with multiple book output formats,
including HTML. Choose one of two HTML book output formats:
}
\details{
\itemize{
\item Use \code{create_bs4_ucsc_course()} to use a \code{bookdown::bs4_book()} output format for course docs
}

The function will create a folder with file structure for a bookdown project,
and example files with information on how to start.
}
