% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc-distill.R
\name{create_ucsc_distill_website}
\alias{create_ucsc_distill_website}
\title{Create a distill fashion course themed for UCSC}
\usage{
create_ucsc_distill_website(dir, title, gh_pages = FALSE, edit = interactive())
}
\arguments{
\item{dir}{absolute path}

\item{title}{website title}

\item{gh_pages}{are you going to use GH pages?, Default: FALSE}

\item{edit}{interactively decide index.Rmc, Default: interactive()}
}
\description{
can easily create new distill website with USCS themeing.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[distill]{character(0)}}
}
