% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc-themes.R
\docType{data}
\name{font_inter}
\alias{font_inter}
\title{Inter font declaration (Secondary)}
\format{
An object of class \code{character} of length 1.
}
\usage{
font_inter
}
\description{
\code{font_intr} == "\code{Inter}"
}
\note{
\code{font_intr} (a.k.a. "\code{Inter}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{internal}
