% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc-themes.R
\docType{data}
\name{font_robo}
\alias{font_robo}
\title{Roboto font declaration (Primary)}
\format{
An object of class \code{character} of length 1.
}
\usage{
font_robo
}
\description{
\code{font_robo_} == "\code{Roboto }"
}
\note{
\code{font_rob} (a.k.a. "\code{Roboto}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{internal}
