% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc-palette.R
\name{ucsc_cols}
\alias{ucsc_cols}
\title{Function to extract ucsc colours as hex codes}
\usage{
ucsc_cols(...)
}
\arguments{
\item{...}{Character names of ucsc_colors}
}
\description{
Function to extract ucsc colours as hex codes
}
