% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{ucsc_font}
\alias{ucsc_font}
\title{ucsc theme fonts}
\usage{
ucsc_font(type = c("body", "title"))
}
\arguments{
\item{type}{body or title font type}
}
\description{
Returns family name of ucsc fonts, to
be used in plots. If for instance used the the
ggplot2-theme functions.
}
