% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucsc-palette.R
\name{ucsc_pal}
\alias{ucsc_pal}
\title{Return function to interpolate a ucsc color palette}
\usage{
ucsc_pal(palette = "redblack", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ucsc_palettes}

\item{reverse}{Logical indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to \code{\link[grDevices]{colorRampPalette}}}
}
\description{
Return function to interpolate a ucsc color palette
}
