% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucscthemes-package.R
\docType{package}
\name{ucscthemes-package}
\alias{ucscthemes}
\alias{ucscthemes-package}
\title{ucscthemes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Niccolò Salvini \email{niccolo.salvini27@gmail.com}

}
\keyword{internal}
