% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucscthemes-package.R
\docType{package}
\name{ucscthemes-package}
\alias{ucscthemes}
\alias{ucscthemes-package}
\title{ucscthemes: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://niccolosalvini.me/ucscthemes/}
}

}
\author{
\strong{Maintainer}: Niccolò Salvini \email{niccolo.salvini27@gmail.com}

}
\keyword{internal}
