% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucscthemes-package.R
\docType{package}
\name{ucscthemes-package}
\alias{ucscthemes}
\alias{ucscthemes-package}
\title{ucscthemes: Rmarkdown templates for UCSC}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of wrappers to quickly start up website, slides and materials branded UCSC
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ucscthemes.netlify.app/index.html}
  \item \url{https://github.com/NiccoloSalvini/ucscthemes}
}

}
\author{
\strong{Maintainer}: Niccolò Salvini \email{niccolo.salvini27@gmail.com}

}
\keyword{internal}
